create db, FIX

+++ src/utils.py
@@ -2,13 +2,13 @@ import psycopg2
 import configparser
 import os
 from psycopg2.extensions import ISOLATION_LEVEL_AUTOCOMMIT
-
+from tabulate import tabulate
 
 def create_or_update_config(file_path):
     config = configparser.ConfigParser()
     print("Создание и настройка файла конфигурации доступа к базе данных.")
     if os.path.exists(file_path):
-        user_choice = input(f"Файл конфигурации {file_path} уже существует. Хотите пересоздать его? (y/n): ")
+        user_choice = input(f"Файл конфигурации {file_path} уже существует. Хотите пересоздать его? (y - да): ")
         if user_choice.lower() != 'y':
             print("Используется текущий файл конфигурации.")
             return
@@ -65,8 +65,15 @@ def create_db(file_path):
 
     if exists:
         # Удаление базы данных, если она существует
-        cur.execute(f"DROP DATABASE {dbname}")
-        print(f"База данных {dbname} удалена.")
+        user_choice = input(f"База данных {dbname} уже существует. Хотите удалить её? (y - да): ")
+        if user_choice.lower() == 'y':
+            cur.execute(f"DROP DATABASE {dbname}")
+            print(f"База данных {dbname} удалена.")
+        else:
+            print("Используется текущая база данных.")
+            cur.close()
+            conn.close()
+            return
 
     # Создание новой базы данных
     cur.execute(f"CREATE DATABASE {dbname}")
@@ -77,42 +84,42 @@ def create_db(file_path):
     conn.close()
 
 
-def create_tables(file_path):
-    config = configparser.ConfigParser()
-    config.read(file_path, encoding='utf-8')
-    dbname = config.get('postgresql', 'dbname')
-    user = config.get('postgresql', 'user')
-    password = config.get('postgresql', 'password')
-    host = config.get('postgresql', 'host')
-    port = config.get('postgresql', 'port')
-
-    conn = psycopg2.connect(dbname=dbname, user=user, password=password, host=host, port=port)
-    cur = conn.cursor()
-
-    # Создание таблиц
-    cur.execute('''
-    CREATE TABLE IF NOT EXISTS employers (
-        employer_id VARCHAR(8) PRIMARY KEY,
-        name VARCHAR(255) NOT NULL,
-        url VARCHAR(255)
-    );
-    ''')
-
-    cur.execute('''
-    CREATE TABLE IF NOT EXISTS vacancies (
-        vacancy_id VARCHAR(50) UNIQUE NOT NULL,
-        name VARCHAR(255) NOT NULL,
-        employer_id VARCHAR(8) NOT NULL,
-        url VARCHAR(255),
-        salary_from INTEGER,
-        salary_to INTEGER,
-        currency VARCHAR(3),
-        FOREIGN KEY (employer_id) REFERENCES employers (employer_id)
-    );
-    ''')
-    conn.commit()
-    cur.close()
-    conn.close()
+# def create_tables(file_path):
+#     config = configparser.ConfigParser()
+#     config.read(file_path, encoding='utf-8')
+#     dbname = config.get('postgresql', 'dbname')
+#     user = config.get('postgresql', 'user')
+#     password = config.get('postgresql', 'password')
+#     host = config.get('postgresql', 'host')
+#     port = config.get('postgresql', 'port')
+#
+#     conn = psycopg2.connect(dbname=dbname, user=user, password=password, host=host, port=port)
+#     cur = conn.cursor()
+#
+#     # Создание таблиц
+#     cur.execute('''
+#     CREATE TABLE IF NOT EXISTS employers (
+#         employer_id VARCHAR(8) PRIMARY KEY,
+#         name VARCHAR(255) NOT NULL,
+#         url VARCHAR(255)
+#     );
+#     ''')
+#
+#     cur.execute('''
+#     CREATE TABLE IF NOT EXISTS vacancies (
+#         vacancy_id VARCHAR(50) UNIQUE NOT NULL,
+#         name VARCHAR(255) NOT NULL,
+#         employer_id VARCHAR(8) NOT NULL,
+#         url VARCHAR(255),
+#         salary_from INTEGER,
+#         salary_to INTEGER,
+#         currency VARCHAR(3),
+#         FOREIGN KEY (employer_id) REFERENCES employers (employer_id)
+#     );
+#     ''')
+#     conn.commit()
+#     cur.close()
+#     conn.close()
 
 
 def draw_progress_bar(current: int, total: int, bar_length: int = 30):
@@ -129,3 +136,4 @@ def draw_progress_bar(current: int, total: int, bar_length: int = 30):
     bar = '[' + '#' * num_ticks + '_' * (bar_length - num_ticks) + ']'
     percentage = int(progress * 100)
     return f'{bar} {percentage}%'
+
